+++ CHANGES.md
@@ -2,3 +2,4 @@ Changes
 =======
 
 - [Mon Jun 22] Added namespace support for brances.
+- [Tue Aug 18] Re code the new parser based on tokens.